Keep OpenAPI config together, and remove old docs generation
