+++ src/config/openapi.ts
@@ -7,7 +7,7 @@ const envSchema = z.object({
 
 const env = envSchema.parse(process.env)
 
-const openAPITags = {
+const openAPITagDefinitions = {
   Companies: {
     description: 'Companies and related resources',
   },
@@ -29,16 +29,29 @@ const openAPITags = {
   },
 } as const
 
-type TagName = keyof typeof openAPITags
-type Tag = (typeof openAPITags)[TagName] & { name: TagName }
+type TagName = keyof typeof openAPITagDefinitions
+type Tag = (typeof openAPITagDefinitions)[TagName] & { name: TagName }
 
-export default {
-  prefix: env.OPENAPI_PREFIX,
-  tags: Object.entries(openAPITags).reduce((tags, [name, tag]) => {
+const openAPITags = Object.entries(openAPITagDefinitions).reduce(
+  (tags, [name, tag]) => {
     const tagName = name as unknown as TagName
     tags[tagName] = { name: tagName, ...tag }
     return tags
-  }, {} as Record<TagName, Tag>),
+  },
+  {} as Record<TagName, Tag>
+)
+
+/**
+ * Format valid OpenAPI tags as an array.
+ */
+export function getTags(...tags: (keyof typeof openAPITags)[]) {
+  return tags
+}
+
+export default {
+  prefix: env.OPENAPI_PREFIX,
+  tags: openAPITags,
+
   title: 'Klimatkollen API Reference',
   description: `
 ![Klimatkollen Logo](https://beta.klimatkollen.se/klimatkollen_logo.svg)
