--- src/openapi/utils.ts
@@ -1,8 +1,0 @@
-import openAPIConfig from '../config/openapi'
-
-/**
- * Format valid OpenAPI tags as an array.
- */
-export function getTags(...tags: (keyof typeof openAPIConfig.tags)[]) {
-  return tags
-}
